backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-multipart>=0.0.9
transformers>=4.43
accelerate>=0.33
bitsandbytes>=0.43
pymupdf>=1.24.0
pillow>=10.0.0
qwen-vl-utils>=0.0.8
